chore: calculate page languages outside of the layout context
